# Doctors/public/swagger.json
{
	"components": {
...
		"schemas": {
			"SaveUpdateResDoctor": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"FullName": {
						"type": "string"
					},
					"Desgination": {
						"type": "string"
					},
					"cell": {
						"type": "string"
					},
					"JoinDate": {
						"type": "string"
					},
					"adress": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"FullName",
					"Desgination",
					"cell",
					"JoinDate",
					"adress",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetDoctor": {
				"properties": {
					"id": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveReqDoctor": {
				"properties": {
					"FullName": {
						"type": "string"
					},
					"Desgination": {
						"type": "string"
					},
					"cell": {
						"type": "string"
					},
					"JoinDate": {
						"type": "string"
					},
					"adress": {
						"type": "string"
					}
				},
				"required": [
					"FullName",
					"Desgination",
					"cell",
					"JoinDate",
					"adress"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateReqDoctor": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"FullName": {
						"type": "string"
					},
					"Desgination": {
						"type": "string"
					},
					"cell": {
						"type": "string"
					},
					"JoinDate": {
						"type": "string"
					},
					"adress": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"FullName",
					"Desgination",
					"cell",
					"JoinDate",
					"adress"
				],
				"type": "object",
				"additionalProperties": false
			},
			"_LeanDocument__LeanDocument_T__": {
				"properties": {},
				"type": "object"
			},
			"Pick__LeanDocument_T_.Exclude_keyof_LeanDocument_T_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested__": {
				"properties": {
					"_id": {
						"$ref": "#/components/schemas/_LeanDocument__LeanDocument_T__",
						"description": "This documents _id."
					},
					"__v": {
						"description": "This documents __v."
					},
					"id": {
						"description": "The string version of this documents _id."
					}
				},
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit__LeanDocument_this_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested_": {
				"$ref": "#/components/schemas/Pick__LeanDocument_T_.Exclude_keyof_LeanDocument_T_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"LeanDocument_this_": {
				"$ref": "#/components/schemas/Omit__LeanDocument_this_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested_"
			},
			"IDoctor": {
				"$ref": "#/components/schemas/LeanDocument_this_"
			},
			"DeleteDoctor": {
				"properties": {
					"id": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RegisterUpdateResDoctor": {
				"properties": {
					"like": {},
					"_id": {
						"type": "string"
					},
					"FirstName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"cell": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"adress": {
						"type": "string"
					},
					"disease": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"like",
					"_id",
					"FirstName",
					"email",
					"cell",
					"password",
					"adress",
					"disease",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RegisterDoctor": {
				"properties": {
					"userName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"userName",
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"loginDoctor": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"like": {
				"properties": {
					"userId": {
						"type": "string"
					},
					"doctorId": {
						"type": "string"
					}
				},
				"required": [
					"userId",
					"doctorId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RegisterUpdateResPatient": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"FullName": {
						"type": "string"
					},
					"resetPasswordToken": {
						"type": "string"
					},
					"resetPasswordExpires": {
						"type": "number",
						"format": "double"
					},
					"email": {
						"type": "number",
						"format": "double"
					},
					"otpCode": {
						"type": "string"
					},
					"cell": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"adress": {
						"type": "string"
					},
					"disease": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"FullName",
					"resetPasswordToken",
					"resetPasswordExpires",
					"email",
					"otpCode",
					"cell",
					"password",
					"adress",
					"disease",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"loginPatient": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"LLinks": {
				"properties": {
					"rel": {
						"type": "string"
					},
					"href": {
						"type": "string"
					}
				},
				"required": [
					"rel",
					"href"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ShowRegion": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"nameKey": {
						"type": "string"
					},
					"code": {
						"type": "string"
					},
					"links": {
						"items": {
							"$ref": "#/components/schemas/LLinks"
						},
						"type": "array"
					}
				},
				"required": [
					"_id",
					"name",
					"nameKey",
					"code",
					"links"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Links3": {
				"properties": {
					"rel": {
						"type": "string"
					},
					"href": {
						"type": "string"
					}
				},
				"required": [
					"rel",
					"href"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResMetaData": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"dataSource": {
						"type": "string"
					},
					"extractDate": {
						"type": "string"
					},
					"itemCount": {
						"type": "string"
					},
					"returncode": {
						"type": "string"
					},
					"totalCount": {
						"type": "string"
					},
					"totalPages": {
						"type": "string"
					},
					"pageSize": {
						"type": "string"
					},
					"pageNumber": {
						"type": "string"
					},
					"links": {
						"items": {
							"$ref": "#/components/schemas/Links3"
						},
						"type": "array"
					}
				},
				"required": [
					"_id",
					"dataSource",
					"extractDate",
					"itemCount",
					"returncode",
					"totalCount",
					"totalPages",
					"pageSize",
					"pageNumber",
					"links"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CombineArray": {
				"properties": {
					"region": {
						"items": {
							"$ref": "#/components/schemas/ShowRegion"
						},
						"type": "array"
					},
					"meta": {
						"items": {
							"$ref": "#/components/schemas/ResMetaData"
						},
						"type": "array"
					}
				},
				"required": [
					"region",
					"meta"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Links": {
				"properties": {
					"rel": {
						"type": "string"
					},
					"href": {
						"type": "string"
					}
				},
				"required": [
					"rel",
					"href"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Region": {
				"properties": {
					"name": {
						"type": "string"
					},
					"nameKey": {
						"type": "string"
					},
					"code": {
						"type": "string"
					},
					"links": {
						"items": {
							"$ref": "#/components/schemas/Links"
						},
						"type": "array"
					}
				},
				"required": [
					"name",
					"nameKey",
					"code",
					"links"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Links222": {
				"properties": {
					"rel": {
						"type": "string"
					},
					"href": {
						"type": "string"
					}
				},
				"required": [
					"rel",
					"href"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReqMetaDataa": {
				"properties": {
					"dataSource": {
						"type": "string"
					},
					"extractDate": {
						"type": "number",
						"format": "double"
					},
					"itemCount": {
						"type": "number",
						"format": "double"
					},
					"returncode": {
						"type": "number",
						"format": "double"
					},
					"totalCount": {
						"type": "number",
						"format": "double"
					},
					"totalPages": {
						"type": "number",
						"format": "double"
					},
					"pageSize": {
						"type": "number",
						"format": "double"
					},
					"pageNumber": {
						"type": "number",
						"format": "double"
					},
					"links": {
						"items": {
							"$ref": "#/components/schemas/Links222"
						},
						"type": "array"
					}
				},
				"required": [
					"dataSource",
					"extractDate",
					"itemCount",
					"returncode",
					"totalCount",
					"totalPages",
					"pageSize",
					"pageNumber",
					"links"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pagination": {
				"properties": {
					"pageNumber": {
						"type": "number",
						"format": "double"
					},
					"pageSize": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"pageNumber",
					"pageSize"
				],
				"type": "object",
				"additionalProperties": false
			},
			"getRegionById": {
				"properties": {
					"id": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"sortingRegions": {
				"properties": {
					"SortBy": {
						"type": "string"
					}
				},
				"required": [
					"SortBy"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ShowImageee": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"image": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"image"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReqImageUpload": {
				"properties": {
					"image": {
						"type": "string"
					}
				},
				"required": [
					"image"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "admin_token",
				"in": "header"
			},
			"jwt": {
				"type": "http",
				"scheme": "Bearer",
				"name": "Authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "swaggerapis2",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/doctor/getDoctor": {
			"post": {
				"operationId": "Getdoctor",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResDoctor"
								}
							}
						}
					}
				},
				"description": "jskhd ksjhdkjsahd skhdjkshdkjs sd sdsdsds",
				"summary": "Get Doctro",
				"tags": [
					"doctor"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GetDoctor"
							}
						}
					}
				}
			}
		},
		"/doctor/savedoctor": {
			"post": {
				"operationId": "Savedoctor",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResDoctor"
								}
							}
						}
					}
				},
				"summary": "Save Doctor",
				"tags": [
					"doctor"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SaveReqDoctor"
							}
						}
					}
				}
			}
		},
		"/doctor/updatedoctor": {
			"put": {
				"operationId": "UpdateDoctor",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResDoctor"
								}
							}
						}
					}
				},
				"tags": [
					"doctor"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateReqDoctor"
							}
						}
					}
				}
			}
		},
		"/doctor/deletedoctor": {
			"delete": {
				"operationId": "Deletdoctor",
				"responses": {
					"200": {
						"description": "Doctor deleted",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IDoctor"
								}
							}
						}
					}
				},
				"tags": [
					"doctor"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeleteDoctor"
							}
						}
					}
				}
			}
		},
		"/doctor/getDoctorlist": {
			"post": {
				"operationId": "GetdoctorList",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/SaveUpdateResDoctor"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"doctor"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/doctor/register": {
			"post": {
				"operationId": "RegisterDoctor",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RegisterUpdateResDoctor"
								}
							}
						}
					}
				},
				"tags": [
					"doctor"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RegisterDoctor"
							}
						}
					}
				}
			}
		},
		"/doctor/loginUser": {
			"post": {
				"operationId": "LoginEmailUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RegisterUpdateResDoctor"
								}
							}
						}
					}
				},
				"tags": [
					"doctor"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/loginDoctor"
							}
						}
					}
				}
			}
		},
		"/doctor/like": {
			"put": {
				"operationId": "FindDoctorById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RegisterUpdateResDoctor"
								}
							}
						}
					}
				},
				"tags": [
					"doctor"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/like"
							}
						}
					}
				}
			}
		},
...
		"/patient/patient-loginUser": {
			"post": {
				"operationId": "LoginEmailUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RegisterUpdateResPatient"
								}
							}
						}
					}
				},
				"tags": [
					"patient"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/loginPatient"
							}
						}
					}
				}
			}
		},
...
		"/region/save-region": {
			"post": {
				"operationId": "SaveData",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ShowRegion"
								}
							}
						}
					}
				},
				"tags": [
					"region"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Region"
							}
						}
					}
				}
			}
		},
		"/region/save-meta": {
			"post": {
				"operationId": "SaveeMetaData",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResMetaData"
								}
							}
						}
					}
				},
				"tags": [
					"region"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ReqMetaDataa"
							}
						}
					}
				}
			}
		},
		"/region/pagination": {
			"post": {
				"operationId": "Pagination",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CombineArray"
								}
							}
						}
					}
				},
				"tags": [
					"region"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Pagination"
							}
						}
					}
				}
			}
		},
		"/region/get-region-by-id": {
			"post": {
				"operationId": "GetRegionBy",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ShowRegion"
								}
							}
						}
					}
				},
				"tags": [
					"region"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/getRegionById"
							}
						}
					}
				}
			}
		},
		"/region/sorting": {
			"post": {
				"operationId": "Sorting",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ShowRegion"
								}
							}
						}
					}
				},
				"tags": [
					"region"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/sortingRegions"
							}
						}
					}
				}
			}
		},
		"/region/upload-image": {
			"post": {
				"operationId": "ImageUp",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ShowImageee"
								}
							}
						}
					}
				},
				"tags": [
					"region"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ReqImageUpload"
							}
						}
					}
				}
			}
		},
		"/region/get-region-by-id/{id}": {
			"post": {
				"operationId": "GetRegionBy2",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ShowRegion"
								}
							}
						}
					}
				},
				"tags": [
					"region"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/region/paginations/region?page={page}&limit={limit}": {
			"get": {
				"operationId": "PaginationQuery",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CombineArray"
								}
							}
						}
					}
				},
				"tags": [
					"region"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "page",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}